mods

+++ external_distr/create_demo_enstore_environment.sh
@@ -82,6 +82,9 @@ then
     echo "will install a minimal enstore configuration file: ${ENSTORE_DIR}/etc/minimal_enstore.conf"
     echo "it can be replased later"
     cp -p ${ENSTORE_DIR}/etc/minimal_enstore.conf $ENSTORE_CONFIG_FILE
+fi
+if [ ! -f /usr/local/etc/setups.sh -a $this_host = $ENSTORE_CONFIG_HOST ];
+    rm -rf /usr/local/etc/setups_rpm.sh
     sed -e "s?e_home=?e_home=$ENSTORE_HOME?" $ENSTORE_DIR/external_distr/setups.sh > /usr/local/etc/setups_rpm.sh
     ln -s /usr/local/etc/setups_rpm.sh /usr/local/etc/setups.sh
     host_name=`uname -n | cut -f1 -d\.`
